When using the Spanner Emulator, sort by column name
Since the value of ORDINAL_POSITION is NULL for the STORING column, the order is undetermined.
Spanner Instances are implicitly returned in the order of their definition, but the Spanner Emulator's specifications make the order random.
Currently, the Spanner Instance's Information Schema and DDL return the results in the order expected by the developer.
For this reason, only when using the Spanner Emulator are we sorted by column name to fix the order.
In reality, using the Spanner Emulator's Information Schema is not recommended, and this is a measure only for Unit Testing.

refs

* https://github.com/cloudspannerecosystem/yo/issues/154